martalp/s3.py
from django.conf import settings
from django.core.files.storage import get_storage_class
from storages.backends.s3boto import S3BotoStorage


class Static(S3BotoStorage):
    location = settings.STATIC


class Media(S3BotoStorage):
    location = settings.MEDIA


class CachedS3BotoStorage(S3BotoStorage):
    location = settings.STATIC
    """
    S3 storage backend that saves the files locally, too.
    """

    def __init__(self, *args, **kwargs):
        super(CachedS3BotoStorage, self).__init__(*args, **kwargs)
        self.local_storage = get_storage_class(
            "compressor.storage.GzipCompressorFileStorage")()

    def save(self, name, content):
        name = super(CachedS3BotoStorage, self).save(name, content)
        self.local_storage._save(name, content)
        return name