requirements.txt
beautifulsoup4==4.4.1
boto==2.39.0
botocore==1.3.23
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-appconf==1.0.1
django-compressor==2.0
django-modelcluster==1.1
django-storages==1.1.8
-e git+https://github.com/caskroom/homebrew-cask.git@89d950bb46e61d77a304ebe07c8137976749ac42#egg=django_storages_redux
django-taggit==0.18.0
django-toolbelt==0.0.1
django-treebeard==3.0
djangorestframework==3.3.2
docutils==0.12
gunicorn==19.4.5
html5lib==0.9999999
jmespath==0.9.0
Pillow==3.1.0
psycopg2==2.6.1
python-dateutil==2.4.2
pytz==2015.7
rcssmin==1.0.6
rjsmin==1.0.12
s3utils==0.5.3
six==1.10.0
static3==0.6.1
Unidecode==0.4.19
wagtail==1.3rc1
wheel==0.26.0
Willow==0.2.2
